- 코드 리팩토링 content 의 review, notice, notification, faq 앱을 분리하기로 함. 분리 작업 완료
