+++ .idea/misc.xml
@@ -7,7 +7,4 @@
   <component name="PythonCompatibilityInspectionAdvertiser">
     <option name="version" value="3" />
   </component>
-
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12.7 WSL (Ubuntu-22.04): (/home/jang/.cache/pypoetry/virtualenvs/tataro-project-jn1SXzOW-py3.12/bin/python)" project-jdk-type="Python SDK" />
-
 </project>
