+++ content/urls.py
@@ -1,15 +1,9 @@
 from django.urls import path
 
 from content.consumers import NotificationConsumer
-from content.routes.faq_router import faq_detail_update_delete, faq_list_or_create
-from content.routes.notice_router import (
-    notice_detail_update_delete,
-    notice_list_or_create,
-)
-from content.routes.review_router import (
-    review_detail_update_delete,
-    review_list_or_create,
-)
+from faq.views import faq_detail_update_delete, faq_list_or_create
+from notice.views import notice_detail_update_delete, notice_list_or_create
+from review.views import review_detail_update_delete, review_list_or_create
 
 urlpatterns = [
     path("notices/", notice_list_or_create, name="notice-list-create"),
