+++ helpers/models.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+
+class BaseModel(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True, verbose_name="생성일")
+    updated_at = models.DateTimeField(auto_now=True, verbose_name="수정일")
+
+    class Meta:
+        abstract = True
